Add files via upload
This repository contains the Python code used to analyze a dataset focused on migraine severity, specifically exploring the relationship between exercise habits and the intensity and frequency of migraines. The project examines data from a group of 52 patients, analyzing key factors such as exercise routines, BMI, age, and migraine severity scores (MIDAS and MSQ). The main objective is to understand how different levels of exercise, especially regularity and duration, correlate with the severity of migraines. Additionally, the dataset includes variables related to patient demographics and other health factors, which are analyzed to identify potential predictors of migraine severity. The analysis involves descriptive statistics, correlation tests, regression analysis, and hypothesis testing to draw meaningful conclusions about the impact of exercise on migraine outcomes. The code is designed to be executed in Google Colab and uses Python libraries such as pandas, seaborn, statsmodels, and matplotlib to process the data, perform statistical analyses, and generate visualizations of the results. This repository includes all relevant code and notebooks for reproducing the analyses and visualizing the findings.